Add a test that the generated file is in sync with the generator.
