+++ tools/wasi-headers/src/main.rs
@@ -1,33 +1,11 @@
-mod c_header;
-
-use crate::c_header::to_c_header;
-use anyhow::{anyhow, Result};
-use std::fs::{read_dir, File};
-use std::io::{self, Write};
-use witx::load;
-
-pub fn generate() -> Result<String> {
-    let inputs = read_dir("WASI/phases/snapshot/witx")?
-        .map(|res| res.map(|e| e.path()))
-        .collect::<Result<Vec<_>, io::Error>>()?;
-
-    // TODO: drop the anyhow! part once witx switches to anyhow.
-    let doc = load(&inputs).map_err(|e| anyhow!(e.to_string()))?;
-
-    let inputs_str = &inputs
-        .iter()
-        .map(|p| p.file_name().unwrap().to_str().unwrap().to_string())
-        .collect::<Vec<_>>()
-        .join(", ");
-
-    Ok(to_c_header(&doc, &inputs_str))
-}
+use anyhow::Result;
+use std::fs::File;
+use std::io::Write;
+use wasi_headers::generate;
 
 pub fn main() -> Result<()> {
-    let c_header = generate();
-    let mut file = File::create("../../libc-bottom-half/headers/public/wasi/api.h")
-        .expect("create output file");
-    file.write_all(c_header.as_bytes())
-        .expect("write output file");
+    let c_header = generate()?;
+    let mut file = File::create("../../libc-bottom-half/headers/public/wasi/api.h")?;
+    file.write_all(c_header.as_bytes())?;
     Ok(())
 }
